requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyQt5==5.15.11
psutil==5.9.5
cryptography==41.0.7
argon2-cffi==23.1.0 